module 3. task final. fix

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        float res = (temperatureCelsius * 9 / 5) + 32;
+        double res = (temperatureCelsius * 9.0 / 5.0) + 32.0;
         System.out.println(res);
     }
 }
